refactor(frontend): 拆分 settings_page 和 info_page 公用组件到 widgets 目录
