+++ lib/pages/info_page.dart
@@ -13,6 +13,7 @@ import 'package:url_launcher/url_launcher.dart';
 
 import '../models/message_item.dart';
 import '../services/sspu_news_service.dart';
+import '../widgets/message_tile.dart';
 import '../services/message_state_service.dart';
 
 /// 信息中心页面
@@ -552,7 +553,7 @@ class _InfoPageState extends State<InfoPage> {
         final message = messages[index];
         final isRead = _stateService.isRead(message.id);
 
-        return _MessageTile(
+        return MessageTile(
           message: message,
           isRead: isRead,
           isDark: isDark,
@@ -595,172 +596,6 @@ class _InfoPageState extends State<InfoPage> {
               : null,
         ),
       ],
-    );
-  }
-}
-
-/// 单条消息列表项组件
-class _MessageTile extends StatelessWidget {
-  final MessageItem message;
-  final bool isRead;
-  final bool isDark;
-  final FluentThemeData theme;
-  final VoidCallback onTap;
-  final VoidCallback onMarkRead;
-
-  const _MessageTile({
-    required this.message,
-    required this.isRead,
-    required this.isDark,
-    required this.theme,
-    required this.onTap,
-    required this.onMarkRead,
-  });
-
-  @override
-  Widget build(BuildContext context) {
-    // 未读消息使用更醒目的样式
-    final titleStyle = isRead
-        ? theme.typography.body?.copyWith(
-            color: theme.resources.textFillColorSecondary,
-          )
-        : theme.typography.bodyStrong;
-
-    return HoverButton(
-      onPressed: onTap,
-      builder: (context, states) {
-        final isHovered = states.isHovered;
-
-        return AnimatedContainer(
-          duration: const Duration(milliseconds: 150),
-          curve: Curves.easeOut,
-          padding: const EdgeInsets.symmetric(vertical: 10, horizontal: 12),
-          decoration: BoxDecoration(
-            color: isHovered
-                ? (isDark
-                    ? Colors.white.withValues(alpha: 0.04)
-                    : Colors.black.withValues(alpha: 0.02))
-                : Colors.transparent,
-            borderRadius: BorderRadius.circular(6),
-          ),
-          child: Row(
-            children: [
-              // 未读指示点
-              Container(
-                width: 8,
-                height: 8,
-                margin: const EdgeInsets.only(right: 10),
-                decoration: BoxDecoration(
-                  shape: BoxShape.circle,
-                  color: isRead ? Colors.transparent : Colors.blue,
-                ),
-              ),
-
-              // 标题 + 标签区域
-              Expanded(
-                child: Column(
-                  crossAxisAlignment: CrossAxisAlignment.start,
-                  children: [
-                    // 消息标题
-                    Text(
-                      message.title,
-                      style: titleStyle,
-                      maxLines: 2,
-                      overflow: TextOverflow.ellipsis,
-                    ),
-                    const SizedBox(height: 6),
-                    // 标签行：tag1 来源类型 + tag2 来源名称 + tag3 内容分类
-                    Wrap(
-                      spacing: 6,
-                      runSpacing: 4,
-                      children: [
-                        _buildTag(
-                          message.sourceType.label,
-                          Colors.blue,
-                        ),
-                        _buildTag(
-                          message.sourceName.label,
-                          Colors.teal,
-                        ),
-                        _buildTag(
-                          message.category.label,
-                          Colors.orange,
-                        ),
-                      ],
-                    ),
-                  ],
-                ),
-              ),
-
-              const SizedBox(width: 12),
-
-              // 右侧：日期 + 操作按钮
-              Column(
-                crossAxisAlignment: CrossAxisAlignment.end,
-                mainAxisSize: MainAxisSize.min,
-                children: [
-                  // 日期
-                  Text(
-                    message.date,
-                    style: theme.typography.caption?.copyWith(
-                      color: theme.resources.textFillColorSecondary,
-                    ),
-                  ),
-                  const SizedBox(height: 6),
-                  // 操作按钮行
-                  Row(
-                    mainAxisSize: MainAxisSize.min,
-                    children: [
-                      // 跳转按钮
-                      Tooltip(
-                        message: '在浏览器中打开',
-                        child: IconButton(
-                          icon: const Icon(
-                            FluentIcons.open_in_new_window,
-                            size: 14,
-                          ),
-                          onPressed: onTap,
-                        ),
-                      ),
-                      const SizedBox(width: 4),
-                      // 标为已读按钮
-                      if (!isRead)
-                        Tooltip(
-                          message: '标为已读',
-                          child: IconButton(
-                            icon: const Icon(
-                              FluentIcons.read,
-                              size: 14,
-                            ),
-                            onPressed: onMarkRead,
-                          ),
-                        ),
-                    ],
-                  ),
-                ],
-              ),
-            ],
-          ),
-        );
-      },
-    );
-  }
-
-  /// 构建标签 badge
-  Widget _buildTag(String text, AccentColor color) {
-    return Container(
-      padding: const EdgeInsets.symmetric(horizontal: 6, vertical: 2),
-      decoration: BoxDecoration(
-        color: color.withValues(alpha: isDark ? 0.2 : 0.1),
-        borderRadius: BorderRadius.circular(4),
-      ),
-      child: Text(
-        text,
-        style: TextStyle(
-          fontSize: 11,
-          color: isDark ? color.lighter : color.dark,
-        ),
-      ),
     );
   }
 }
